feat: :sparkles:学校のクラステーブルの学年カラムのデータ型を変更

+++ api/database/migrations/2023_06_21_123944_create_school_classes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('school_classes', function (Blueprint $table) {
             $table->id();
             $table->string('course');
-            $table->integer('grade');
+            $table->unsignedTinyInteger('grade');
             $table->timestamps();
         });
     }
